Abstractions (#2)
* Start adding some basic graph functions
Things like getting nodes and edged by id, following edges a single step.
Test are currently not passing

* Refactored Edges
Instead of two types of edges, directionality is a parameter.
Edges now take EdgeProperties instead of Map[String, Any] which is nicer.

* Refactored Edges
Instead of two types of edges, directionality is a parameter.
Edges now take EdgeProperties instead of Map[String, Any] which is nicer.

* Cleanup

* Edges - Tests, Generic-ish EdgeProperties, get value by name for edges.
Still not an ideal apprach to properties, and need to do the same work for Nodes.
The failing Edge test shows what's not so hot about this approach

* Add node and edge labels

* Started a refactor to abstract into a graph trait. Following a 'final tagless' approach

* Fix for get in/out edges. Parent/child/sibling will need some thinking

* gitignore ensime

* Basic graph algebra and tests passing, with a simple InMemoryGraph implimentation

We can now build a graph with the builder and traverse it with the basic methods. Implimenting a new graph only requires a few methods, the algebra takes care of the rest.

* Start adding some basic graph functions
Things like getting nodes and edged by id, following edges a single step.
Test are currently not passing

* Basic graph algebra and tests passing, with a simple InMemoryGraph implimentation

We can now build a graph with the builder and traverse it with the basic methods. Implimenting a new graph only requires a few methods, the algebra takes care of the rest.