Added a Transition Graph implementation. Better build. 100% line coverage.
A transition graph is essentially a state machine diagram and maybe that's a
better name for this class. It will be used to run the state transition
simulations that are the core of the sales funnel predictions.

Some other things got rolled into this commit as well:
- An improved build.sbt to help keep the project tidy
  - Mostly this means additional compiler flags for
things like fail on warn
  - And also some nice functional features like better
higher kinded types and implicit conversions
- Updated README
- And 100% line coverage.