Refactor GraphAlg so that generic methods don't refer to the nodes or edges vals
The reason being, we want those to be more flexible. Maybe they are not sets, but instead some collection returned from a db query.
The idea is also that we can provide those vals with a mix-in or through composition or something to be able to add constraints (like only allowing directed edges) to the nodes or edges