Core 3/graph simulation (#5)
* Added a Transition Graph implementation. Better build. 100% line coverage.

A transition graph is essentially a state machine diagram and maybe that's a
better name for this class. It will be used to run the state transition
simulations that are the core of the sales funnel predictions.

Some other things got rolled into this commit as well:
- An improved build.sbt to help keep the project tidy
  - Mostly this means additional compiler flags for
things like fail on warn
  - And also some nice functional features like better
higher kinded types and implicit conversions
- Updated README
- And 100% line coverage.

* Transitions graph. Somehow not included in previous commit

* Update README

* Update README

* Update and rename README to README.md

* Split packagae declaration into two lines

* Refactor GraphAlg so that generic methods don't refer to the nodes or edges vals

The reason being, we want those to be more flexible. Maybe they are not sets, but instead some collection returned from a db query.
The idea is also that we can provide those vals with a mix-in or through composition or something to be able to add constraints (like only allowing directed edges) to the nodes or edges

* Comment formatting

* Graphs are case classes extending an Algebra with a backend as a parameter.
For convenience, we can make the backend nodes, edges, and methods available directly from the graph case class
using implicit conversions that live in the package object in graph/package.scala

* Update README.md

* The collection used for nodes and edged is fully abstracted away.